Refactor App.tsx, extract components, and simplify application flow
