+++ src/pages/AboutPage.tsx
@@ -1,0 +1,90 @@
+const AboutPage = () => {
+  return (
+    <div className="bg-white min-h-screen">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-20">
+        <div className="text-center mb-16">
+          <div className="inline-flex items-center px-6 py-2 bg-gradient-to-r from-green-100 to-blue-100 rounded-full border border-green-200 mb-4">
+            <span className="text-green-800 font-semibold text-sm">🌍 International Organization • African Vocation</span>
+          </div>
+          <h1 className="text-4xl sm:text-5xl font-black text-gray-900 mb-4">
+            About the <span className="text-transparent bg-clip-text bg-gradient-to-r from-green-600 to-blue-600">OSS</span>
+          </h1>
+          <div className="w-24 h-1 bg-gradient-to-r from-green-500 to-blue-500 mx-auto rounded-full"></div>
+        </div>
+        
+        <div className="lg:flex lg:items-start lg:gap-16">
+          <div className="lg:w-1/2 mb-12 lg:mb-0">
+            <div className="bg-white/80 backdrop-blur-sm rounded-2xl p-8 shadow-xl border border-green-100">
+              <p className="text-lg text-gray-700 mb-6 leading-relaxed">
+                The <strong className="text-green-600">Sahara and Sahel Observatory (OSS)</strong> is an international organization with an African vocation, founded in 1992 and based in Tunis since 2000. It mainly works on creating and supporting partnerships to jointly address the challenges related to water resources management, as well as the implementation of international agreements on land degradation, biodiversity and climate change in Africa.
+              </p>
+              <p className="text-lg text-gray-700 mb-6 leading-relaxed">
+                Today, the OSS has <strong className="text-green-600">28 African and 7 non-African member countries</strong>. In addition, the Organization collaborates with 12 entities representatives of West, East and North Africa, as well as several UN agencies and non-governmental Organizations.
+              </p>
+              <p className="text-lg text-gray-700 mb-6 leading-relaxed">
+                Our mission is to help African member countries sustainably manage their natural resources, with focus on <strong className="text-green-600">arid, semi-arid and dry sub-humid areas of Africa</strong> in a particularly disadvantageous climate change context.
+              </p>
+              <p className="text-lg text-gray-700 leading-relaxed">
+                OSS develops concepts and methodologies for <strong className="text-green-600">environmental monitoring, natural resources management and climate change adaptation</strong>, based on our four scientific programs: <strong className="text-blue-600">Land, Water, Climate, and Biodiversity</strong>.
+              </p>
+            </div>
+          </div>
+          <div className="lg:w-1/2 grid grid-cols-1 sm:grid-cols-2 gap-6">
+            <div className="bg-gradient-to-br from-green-50 to-emerald-50 p-8 rounded-2xl border border-green-200 shadow-lg hover:shadow-xl transition-all duration-300 transform hover:scale-105">
+              <div className="w-12 h-12 bg-gradient-to-br from-green-500 to-emerald-600 rounded-xl flex items-center justify-center mb-4">
+                <svg className="w-6 h-6 text-white" fill="currentColor" viewBox="0 0 20 20">
+                  <path fillRule="evenodd" d="M6.267 3.455a3.066 3.066 0 001.745-.723 3.066 3.066 0 013.976 0 3.066 3.066 0 001.745.723 3.066 3.066 0 012.812 2.812c.051.643.304 1.254.723 1.745a3.066 3.066 0 010 3.976 3.066 3.066 0 00-.723 1.745 3.066 3.066 0 01-2.812 2.812 3.066 3.066 0 00-1.745.723 3.066 3.066 0 01-3.976 0 3.066 3.066 0 00-1.745-.723 3.066 3.066 0 01-2.812-2.812 3.066 3.066 0 00-.723-1.745 3.066 3.066 0 010-3.976 3.066 3.066 0 00.723-1.745 3.066 3.066 0 012.812-2.812zm7.44 5.252a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+                </svg>
+              </div>
+              <h3 className="text-xl font-bold text-green-800 mb-4">Our Mission</h3>
+              <ul className="space-y-3 text-green-700">
+                <li className="flex items-start">
+                  <svg className="h-5 w-5 text-green-500 mr-3 mt-0.5 flex-shrink-0" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+                  </svg>
+                  <span className="font-medium">Implementation of multilateral agreements on land degradation, biodiversity and climate change</span>
+                </li>
+                <li className="flex items-start">
+                  <svg className="h-5 w-5 text-green-500 mr-3 mt-0.5 flex-shrink-0" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+                  </svg>
+                  <span className="font-medium">Promotion of regional and international initiatives addressing environmental challenges</span>
+                </li>
+                <li className="flex items-start">
+                  <svg className="h-5 w-5 text-green-500 mr-3 mt-0.5 flex-shrink-0" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+                  </svg>
+                  <span className="font-medium">Alignment of approaches and unification of methodologies for sustainable land and water management</span>
+                </li>
+              </ul>
+            </div>
+            <div className="bg-gradient-to-br from-blue-50 to-cyan-50 p-8 rounded-2xl border border-blue-200 shadow-lg hover:shadow-xl transition-all duration-300 transform hover:scale-105">
+              <div className="w-12 h-12 bg-gradient-to-br from-blue-500 to-cyan-600 rounded-xl flex items-center justify-center mb-4">
+                <svg className="w-6 h-6 text-white" fill="currentColor" viewBox="0 0 20 20">
+                  <path d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
+                </svg>
+              </div>
+              <h3 className="text-xl font-bold text-blue-800 mb-4">Our Impact</h3>
+              <ul className="space-y-3 text-blue-700">
+                <li className="flex items-start">
+                  <svg className="h-5 w-5 text-blue-500 mr-3 mt-0.5 flex-shrink-0" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+                  </svg>
+                  <span className="font-medium">Accredited by Green Climate Fund & Adaptation Fund</span>
+                </li>
+                <li className="flex items-start">
+                  <svg className="h-5 w-5 text-blue-500 mr-3 mt-0.5 flex-shrink-0" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" />
+                  </svg>
+                  <span className="font-medium">12+ regional partnerships across Africa</span>
+                </li>
+              </ul>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default AboutPage;
