+++ src/pages/AdminDashboard.tsx
@@ -1,0 +1,301 @@
+import { useEffect, useState } from 'react';
+import type { Log, User } from '../types';
+import UserForm from '../components/forms/UserForm';
+
+const AdminDashboard = () => {
+  const [users, setUsers] = useState<User[]>([]);
+  const [logs, setLogs] = useState<Log[]>([]);
+  const [showUserForm, setShowUserForm] = useState(false);
+  const [editingUser, setEditingUser] = useState<User | null>(null);
+  const [activeTab, setActiveTab] = useState<'users' | 'logs'>('users');
+  const [isLoading, setIsLoading] = useState(true);
+  const [error] = useState('');
+  
+  useEffect(() => {
+    const load = async () => {
+      await Promise.all([fetchUsers(), fetchLogs()]);
+      setIsLoading(false);
+    };
+    load();
+  }, []);
+  
+  const fetchUsers = async () => {
+    const token = localStorage.getItem('token');
+    const res = await fetch('http://localhost:8000/users', { headers: { 'Authorization': `Bearer ${token}` } });
+    const data = await res.json();
+    setUsers(data);
+  };
+  
+  const fetchLogs = async () => {
+    const token = localStorage.getItem('token');
+    const res = await fetch('http://localhost:8000/logs', { headers: { 'Authorization': `Bearer ${token}` } });
+    const data = await res.json();
+    setLogs(data);
+  };
+  
+  const handleSaveUser = async (userData: any) => {
+    const token = localStorage.getItem('token');
+    const url = editingUser ? `http://localhost:8000/users/${editingUser.id}` : 'http://localhost:8000/users';
+    const method = editingUser ? 'PUT' : 'POST';
+    const res = await fetch(url, {
+      method,
+      headers: { 'Content-Type': 'application/json', 'Authorization': `Bearer ${token}` },
+      body: JSON.stringify(userData),
+    });
+    if (res.ok) {
+      await fetchUsers();
+      setShowUserForm(false);
+      setEditingUser(null);
+    }
+  };
+  
+  const handleDeleteUser = async (id: number) => {
+    if (window.confirm('Delete this user?')) {
+      const token = localStorage.getItem('token');
+      const res = await fetch(`http://localhost:8000/users/${id}`, {
+        method: 'DELETE',
+        headers: { 'Authorization': `Bearer ${token}` },
+      });
+      if (res.ok) await fetchUsers();
+    }
+  };
+  
+  if (isLoading) return <div className="flex justify-center items-center h-64"><div className="animate-spin h-12 w-12 border-t-2 border-green-600"></div></div>;
+  
+  return (
+    <div className="py-8 bg-gray-50 min-h-screen">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="mb-8">
+          <nav className="flex space-x-8 border-b border-gray-200 w-full">
+            <button 
+              onClick={() => setActiveTab('users')} 
+              className={`py-4 px-2 border-b-2 font-medium text-sm transition-colors duration-200 ${
+                activeTab === 'users' 
+                  ? 'border-green-500 text-green-600' 
+                  : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+              }`}
+            >
+              <div className="flex items-center space-x-2">
+                <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197m13.5-9a2.5 2.5 0 11-5 0 2.5 2.5 0 015 0z" />
+                </svg>
+                <span>Manage Users</span>
+              </div>
+            </button>
+            <button 
+              onClick={() => setActiveTab('logs')} 
+              className={`py-4 px-2 border-b-2 font-medium text-sm transition-colors duration-200 ${
+                activeTab === 'logs' 
+                  ? 'border-green-500 text-green-600' 
+                  : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+              }`}
+            >
+              <div className="flex items-center space-x-2">
+                <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
+                </svg>
+                <span>System Logs</span>
+              </div>
+            </button>
+          </nav>
+        </div>
+        
+        {error && (
+          <div className="mb-6 p-4 bg-red-50 border border-red-200 rounded-lg">
+            <div className="flex items-center">
+              <svg className="w-5 h-5 text-red-400 mr-3" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+              </svg>
+              <p className="text-red-700 font-medium">{error}</p>
+            </div>
+          </div>
+        )}
+        
+        {activeTab === 'users' && (
+          <div className="space-y-6">
+            <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4">
+              <div>
+                <h2 className="text-3xl font-bold text-gray-900">Manage Users</h2>
+                <p className="text-gray-600 mt-1">Create, edit, and manage system users</p>
+              </div>
+              <button 
+                onClick={() => { setEditingUser(null); setShowUserForm(true); }} 
+                className="inline-flex items-center px-6 py-3 bg-green-600 text-white font-medium rounded-lg shadow-sm hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2 transition-colors duration-200"
+              >
+                <svg className="w-5 h-5 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 6v6m0 0v6m0-6h6m-6 0H6" />
+                </svg>
+                Create User
+              </button>
+            </div>
+            
+            {showUserForm && (
+              <div className="bg-white rounded-xl shadow-lg border border-gray-200 overflow-hidden">
+                <div className="px-6 py-4 bg-gray-50 border-b border-gray-200">
+                  <h3 className="text-lg font-semibold text-gray-900">
+                    {editingUser ? 'Edit User' : 'Create New User'}
+                  </h3>
+                </div>
+                <div className="p-6">
+                  <UserForm
+                    user={editingUser || undefined}
+                    onSave={handleSaveUser}
+                    onCancel={() => { setShowUserForm(false); setEditingUser(null); }}
+                  />
+                </div>
+              </div>
+            )}
+            
+            <div className="bg-white rounded-xl shadow-sm border border-gray-200 overflow-hidden">
+              {users.length === 0 ? (
+                <div className="text-center py-12">
+                  <div className="w-16 h-16 mx-auto mb-4 bg-gray-100 rounded-full flex items-center justify-center">
+                    <svg className="w-8 h-8 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197m13.5-9a2.5 2.5 0 11-5 0 2.5 2.5 0 015 0z" />
+                    </svg>
+                  </div>
+                  <h3 className="text-lg font-medium text-gray-900 mb-2">No users found</h3>
+                  <p className="text-gray-500 mb-6">Get started by creating your first user</p>
+                  <button 
+                    onClick={() => { setEditingUser(null); setShowUserForm(true); }} 
+                    className="inline-flex items-center px-4 py-2 bg-green-600 text-white text-sm font-medium rounded-lg hover:bg-green-700 transition-colors duration-200"
+                  >
+                    <svg className="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 6v6m0 0v6m0-6h6m-6 0H6" />
+                    </svg>
+                    Create User
+                  </button>
+                </div>
+              ) : (
+                <ul className="divide-y divide-gray-100">
+                  {users.map(user => (
+                    <li key={user.id} className="px-6 py-5 hover:bg-gray-50 transition-colors duration-150">
+                      <div className="flex justify-between items-center">
+                        <div className="flex items-center flex-1">
+                          <div className="w-12 h-12 bg-gradient-to-r from-blue-500 to-purple-600 rounded-full flex items-center justify-center mr-4">
+                            <span className="text-white font-semibold text-lg">
+                              {user.name.split(' ').map(n => n[0]).join('').slice(0, 2).toUpperCase()}
+                            </span>
+                          </div>
+                          <div className="flex-1">
+                            <div className="flex items-center gap-3 mb-1">
+                              <h3 className="font-semibold text-gray-900 text-lg">{user.name}</h3>
+                              <span className={`inline-flex items-center px-3 py-1 rounded-full text-xs font-medium ${
+                                user.role === 'admin' 
+                                  ? 'bg-purple-100 text-purple-800' 
+                                  : 'bg-blue-100 text-blue-800'
+                              }`}>
+                                <svg className="w-3 h-3 mr-1" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d={
+                                    user.role === 'admin' 
+                                      ? "M9 12l2 2 4-4m5.618-4.016A11.955 11.955 0 0112 2.944a11.955 11.955 0 01-8.618 3.04A12.02 12.02 0 003 9c0 5.591 3.824 10.29 9 11.622 5.176-1.332 9-6.03 9-11.622 0-1.042-.133-2.052-.382-3.016z"
+                                      : "M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"
+                                  } />
+                                </svg>
+                                {user.role === 'admin' ? 'Administrator' : 'HR Manager'}
+                              </span>
+                            </div>
+                            <div className="flex items-center text-sm text-gray-600">
+                              <svg className="w-4 h-4 mr-2 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M3 8l7.89 7.89a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z" />
+                              </svg>
+                              {user.email}
+                            </div>
+                          </div>
+                        </div>
+                        <div className="flex items-center space-x-3 ml-4">
+                          <button 
+                            onClick={() => { setEditingUser(user); setShowUserForm(true); }} 
+                            className="inline-flex items-center px-4 py-2 border border-gray-300 rounded-lg shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 hover:border-gray-400 focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-green-500 transition-all duration-200"
+                          >
+                            <svg className="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z" />
+                            </svg>
+                            Edit
+                          </button>
+                          <button 
+                            onClick={() => handleDeleteUser(user.id)} 
+                            className="inline-flex items-center px-4 py-2 border border-red-300 rounded-lg shadow-sm text-sm font-medium text-red-700 bg-red-50 hover:bg-red-100 hover:border-red-400 focus:outline-none focus:ring-2 focus:ring-red-500 focus:border-red-500 transition-all duration-200"
+                          >
+                            <svg className="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1-1H8a1 1 0 00-1 1v3M4 7h16" />
+                            </svg>
+                            Delete
+                          </button>
+                        </div>
+                      </div>
+                    </li>
+                  ))}
+                </ul>
+              )}
+            </div>
+          </div>
+        )}
+        
+        {activeTab === 'logs' && (
+          <div className="space-y-6">
+            <div>
+              <h2 className="text-3xl font-bold text-gray-900">System Logs</h2>
+              <p className="text-gray-600 mt-1">Monitor system activity and events</p>
+            </div>
+            
+            {logs.length === 0 ? (
+              <div className="bg-white rounded-xl shadow-sm border border-gray-200">
+                <div className="text-center py-12">
+                  <div className="w-16 h-16 mx-auto mb-4 bg-gray-100 rounded-full flex items-center justify-center">
+                    <svg className="w-8 h-8 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
+                    </svg>
+                  </div>
+                  <h3 className="text-lg font-medium text-gray-900 mb-2">No logs available</h3>
+                  <p className="text-gray-500">System logs will appear here when activities occur</p>
+                </div>
+              </div>
+            ) : (
+              <div className="bg-white rounded-xl shadow-sm border border-gray-200 overflow-hidden">
+                <div className="px-6 py-4 bg-gray-50 border-b border-gray-200">
+                  <div className="flex items-center justify-between">
+                    <h3 className="text-lg font-semibold text-gray-900">Recent Activity</h3>
+                    <div className="flex items-center text-sm text-gray-500">
+                      <svg className="w-4 h-4 mr-1" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
+                      </svg>
+                      Last updated: {new Date().toLocaleString()}
+                    </div>
+                  </div>
+                </div>
+                <ul className="divide-y divide-gray-100 max-h-96 overflow-y-auto">
+                  {logs.map(log => (
+                    <li key={log.id} className="px-6 py-4 hover:bg-gray-50 transition-colors duration-150">
+                      <div className="flex items-start space-x-3">
+                        <div className="flex-shrink-0 w-2 h-2 bg-green-400 rounded-full mt-2"></div>
+                        <div className="flex-1 min-w-0">
+                          <p className="text-gray-900 font-medium leading-relaxed">{log.message}</p>
+                          <div className="flex items-center mt-2 text-sm text-gray-500">
+                            <svg className="w-4 h-4 mr-1" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
+                            </svg>
+                            {new Date(log.created_at).toLocaleString('en-US', {
+                              year: 'numeric',
+                              month: 'short',
+                              day: 'numeric',
+                              hour: '2-digit',
+                              minute: '2-digit',
+                              second: '2-digit'
+                            })}
+                          </div>
+                        </div>
+                      </div>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
+          </div>
+        )}
+      </div>
+    </div>
+  );
+};
+
+export default AdminDashboard;
